Rename registry-related classes and update proxy types for mirror registries
