+++ src/KSail.Models/MirrorRegistry/KSailMirrorRegistry.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel;
+using KSail.Models.Registry;
 
-namespace KSail.Models.Registry;
+namespace KSail.Models.MirrorRegistry;
 
 /// <summary>
 /// A registry to create for the KSail cluster to reconcile flux artifacts, and to proxy and cache images.
@@ -11,7 +12,7 @@ public class KSailMirrorRegistry : KSailRegistry
   /// An optional proxy for the registry to use to proxy and cache images.
   /// </summary>
   [Description("A proxy for the registry to use to proxy and cache images.")]
-  public required KSailRegistryProxy Proxy { get; set; }
+  public required KSailMirrorRegistryProxy Proxy { get; set; }
 
   /// <summary>
   /// The host port of the registry.
